move function to proper place
