+++ napari/layers/labels/_labels_utils.py
@@ -1,7 +1,6 @@
 from functools import lru_cache
 from typing import Tuple
 
-import numba
 import numpy as np
 from scipy import ndimage as ndi
 
@@ -251,65 +250,3 @@ def expand_slice(
         )
         for s, max_size in zip(axes_slice, shape)
     )
-
-
-# def cast_
-
-
-def cast_labels_to_minimum_type_auto(
-    data: np.ndarray, num_colors: int
-) -> np.ndarray:
-    """Perform modulo operation based on number of colors
-
-    Parameters
-    ----------
-    data : np.ndarray
-        Labels data to be casted.
-    num_colors : int
-        Number of unique colors in the data.
-
-    Returns
-    -------
-    np.ndarray
-        Casted labels data.
-    """
-    dtype = minimum_dtype_for_labels(num_colors + 1)
-
-    return _cast_labels_to_minimum_type_auto(data, num_colors + 1, dtype)
-
-
-@numba.njit(parallel=True)
-def _cast_labels_to_minimum_type_auto(
-    data: np.ndarray, num_colors: int, dtype
-) -> np.ndarray:
-    result_array = np.zeros_like(data, dtype=dtype)
-
-    # iterate over data and calculate modulo num_colors assigning to result_array
-
-    for i in numba.prange(data.size):
-        if num_colors > data.flat[i] >= 0:
-            result_array.flat[i] = data.flat[i]
-        else:
-            result_array.flat[i] = data.flat[i] % (num_colors - 1) + 1
-
-    return result_array
-
-
-def minimum_dtype_for_labels(num_colors: int) -> np.dtype:
-    """Return the minimum dtype that can hold the number of colors.
-
-    Parameters
-    ----------
-    num_colors : int
-        Number of unique colors in the data.
-
-    Returns
-    -------
-    np.dtype
-        Minimum dtype that can hold the number of colors.
-    """
-    if num_colors <= np.iinfo(np.uint8).max:
-        return np.dtype(np.uint8)
-    if num_colors <= np.iinfo(np.uint16).max:
-        return np.dtype(np.uint16)
-    return np.dtype(np.float32)
